# Performance: Memoize entity encoding