A good chunk of serverside rendering is spent escaping strings. There is a high likelyness, that we'll see the same strings over and over again. By memoizing those computations we get a nice speedup. A tiny bit more (at least in benchmarks) can be achieved by hoisting out regex creation.

| Benchmark | before (ops/sec) | after (ops/sec) |
|---|---|---|
| Text | 260.183 | **401.336** |
| SearchResults | 637.04 | **730.82** |